Columns max length/char length/display size set to long
Former-commit-id: 5bae4701877ad98215a44621b005cd8dc11cc022

+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/model/exec/DBCColumnMetaData.java
@@ -19,7 +19,7 @@ public interface DBCColumnMetaData extends DBSColumnBase
 {
     int getIndex();
 
-    int getDisplaySize();
+    long getDisplaySize();
 
     String getLabel();
