+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/model/impl/jdbc/struct/JDBCColumn.java
@@ -17,7 +17,7 @@ public abstract class JDBCColumn extends AbstractColumn implements IObjectImageP
     {
     }
 
-    protected JDBCColumn(String name, String typeName, int valueType, int ordinalPosition, int maxLength, int scale,
+    protected JDBCColumn(String name, String typeName, int valueType, int ordinalPosition, long maxLength, int scale,
                          int radix, int precision, boolean nullable, String description)
     {
         super(name, typeName, valueType, ordinalPosition, maxLength, scale, radix, precision, nullable, description);
